backend/requirements.txt
fastapi==0.104.1            
uvicorn==0.23.2             
sqlalchemy==2.0.22          
pydantic==2.4.2             
python-jose==3.3.0          
passlib==1.7.4              
python-multipart==0.0.6     
easyocr==1.7.1              
pymupdf==1.26.0            
langdetect==1.0.9           
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0               
Pillow==10.0.1              
python-dotenv==1.0.0        
bcrypt==4.0.1               
email-validator==2.0.0      
alembic==1.12.0             
coqui-tts==0.26.2
google-generativeai==0.8.5

# pip install coqui-tts
# python -m pip install --upgrade pip
# pip install -r requirements.txt
# pip install python-dotenv